Add separate file for Pacific Time timezone
